Set deleteDeployment force to True
When deleteDeployment force is set to false, a new deployment will
be created to revert the current settings. However, this will a new
deployment, even though the running container has the same config
of the rollback.

Furthermore, it actually kills the running container, so there's
downtime.

By setting force=true, the deployment is canceled, but not new
deployment happens.